requirements.txt

langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
faiss-cpu
python-dotenv
tiktoken
